[FEAT] improve components about user form (#277)
