+++ src/components/LoginForm/index.js
@@ -161,7 +161,7 @@ const LoginFormUI = (props) => {
               ref={register({
                 required: t('VALIDATION_ERROR_PASSWORD_REQUIRED', 'The field Password is required').replace('_attribute_', t('PASSWORD', 'Password')),
                 minLength: {
-                  value: 5,
+                  value: 8,
                   message: t('VALIDATION_ERROR_PASSWORD_MIN_STRING', 'The Password must be at least 8 characters.')
                     .replace('_attribute_', t('PASSWORD', 'Password'))
                     .replace('_min_', 8)
