+++ src/components/UserDetails/styles.js
@@ -34,66 +34,3 @@ export const SideForm = styled.div`
     align-items: center;
   }
 `
-
-export const FormInput = styled.form`
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  width: 100%;
-  margin: 0 auto;
-  div.phone_number {
-    margin-bottom: 20px;
-    width: 100%;
-  }
-  input.form {
-    padding: 10px 15px;
-    width: 100%;
-    margin-bottom: 20px;
-    box-sizing: border-box;
-    &:disabled {
-      background-color: rgba(239, 239, 239, 0.3);
-      cursor: not-allowed;
-    }
-  }
-  button {
-    width: 100%;
-    padding: 7px 0;
-  }
-  @media (min-width: 769px) {
-    flex-wrap: wrap;
-    flex-direction: row;
-    justify-content: space-between;
-    input.form {
-      width: 49%;
-    }
-  }
-`
-
-export const SkeletonForm = styled.div`
-  width: 100%;
-  justify-content: center;
-  height: auto;
-  display: inline-flex;
-  flex-wrap: wrap;
-  padding: 0;
-  margin-top: 0px;
-  > * {
-    width: 80%;
-    margin: 10px 0;
-    display: inline;
-    height: 50px;
-    span{
-      border-radius: 16px;
-      width: 100%;
-      height: 100%;
-    }
-  }
-  @media (min-width: 768px){
-    justify-content: flex-start;
-    > * {
-      height: 20px;
-      width: 35%;
-      margin: 5px
-    }
-  }
-`
