+++ src/components/UserFormDetails/styles.js
@@ -1,0 +1,69 @@
+import styled, { css } from 'styled-components'
+
+export const FormInput = styled.form`
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  width: 100%;
+  margin: 0 auto;
+  div.phone_number {
+    margin-bottom: 20px;
+    width: 100%;
+  }
+  input.form {
+    padding: 10px 15px;
+    width: 100%;
+    margin-bottom: 20px;
+    box-sizing: border-box;
+    &:disabled {
+      background-color: rgba(239, 239, 239, 0.3);
+      cursor: not-allowed;
+    }
+  }
+  button {
+    width: 100%;
+    padding: 7px 0;
+  }
+  @media (min-width: 769px) {
+    flex-wrap: wrap;
+    flex-direction: row;
+    justify-content: space-between;
+    input.form {
+      width: 49%;
+    }
+  }
+`
+
+export const ActionsForm = styled.div`
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  width: 100%;
+
+  button {
+    flex: 1;
+    margin-top: 10px;
+    padding: 7px 0px;
+    width: 49%;
+    box-sizing: border-box;
+
+    &:disabled {
+      cursor: not-allowed;
+    }
+
+    &:nth-child(2) {
+      margin-left: 10px;
+    }
+  }
+
+  ${props => props.theme?.rtl && css`
+    button {
+      &:nth-child(1) {
+        margin-right: 0px;
+      }
+      &:last-child {
+        margin-right: 5px;
+      }
+    }
+  `}
+`
